Added sour's changes + linted them.
